refactor pytest fixtures to conftest.py. add adocstrings.
